formRecogFunction/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions

azure-ai-formrecognizer==3.1.0b3
azure-cosmosdb-table
dataframe_image
pandas
geopy
